版本号 v1.5.0.rc1 外设框架完整实现

+++ include/hw/express-gpu/glv3_decl.h
@@ -404,8 +404,8 @@ typedef struct DebugMessageDesc {
 } __attribute__((packed, aligned(4))) DebugMessageDesc;
 
 typedef struct RingBufferDesc {
-    size_t block_count;
-    size_t block_size;
+    int block_count;
+    int block_size;
     int read_idx;
     int write_idx;
 } __attribute__((packed, aligned(4))) RingBufferDesc;
